refactor: migrate 12 files off legacy utils.js shim, delete the file
All imports of API functions and date utilities that went through the
backwards-compat re-export shim (client/src/utils.js) now point directly
to their canonical modules:

  API functions  → client/src/api/services/<domain>Api.js
  Date utilities → client/src/utils/dateUtils.js

Files updated: contactHandlerUtils.js, MsgOfContactTable.jsx,
MessagesTable.jsx, Review.jsx, useUserActions.js, useCarActions.js,
useMessageActions.js, useServiceActions.js, useReviewsData.js,
useReviewForm.js, useContactForm.js, useAccountActions.js

utils.js is deleted; lint confirms zero remaining imports of the shim.

Closes #161
