+++ client/src/components/landing/reviews/hooks/useReviewForm.js
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef } from "react";
-import { createReview } from "../../../../utils";
+import { createReview } from "../../../../api/services/reviewApi";
 
 /**
  * Custom hook for managing review form
